+++ xen/include/asm-x86/acpi.h
@@ -151,7 +151,7 @@ struct acpi_sleep_info {
 #endif /* CONFIG_ACPI_SLEEP */
 
 #define MAX_MADT_ENTRIES	256
-extern u8 x86_acpiid_to_apicid[];
+extern u32 x86_acpiid_to_apicid[];
 #define MAX_LOCAL_APIC 256
 
 extern u32 pmtmr_ioport;
